Modularize the attribution control code
